レンダリングする関数を別ファイルに追い出す
